--- example.env
@@ -1,7 +1,0 @@
-# Spring Boot Configuration
-SERVER_PORT=8080
-
-# Database Configuration
-SPRING_DATASOURCE_URL=jdbc:mysql://localhost:3306/mydatabase
-SPRING_DATASOURCE_USERNAME=root
-SPRING_DATASOURCE_PASSWORD=password
