+++ src/main/java/com/app/server/domain/BaseTimeEntity.java
@@ -4,6 +4,7 @@ import jakarta.persistence.Column;
 import jakarta.persistence.EntityListeners;
 import jakarta.persistence.MappedSuperclass;
 import lombok.Getter;
+import lombok.ToString;
 import org.springframework.data.annotation.CreatedBy;
 import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.annotation.LastModifiedBy;
@@ -13,9 +14,10 @@ import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 import java.time.LocalDateTime;
 
 @Getter
+@ToString
 @MappedSuperclass // "나를 상속받는 자식한테 내 필드들만 물려줄게. 나 자신은 테이블 만들지 마" 라는 뜻.
 @EntityListeners(AuditingEntityListener.class) // 생성/수정 시간을 자동으로 관리해주는 리스너
-public abstract class BaseEntity {
+public abstract class BaseTimeEntity {
 
     @CreatedDate
     @Column(nullable = false, updatable = false)
