+++ src/main/java/com/app/server/security/JwtUtil.java
@@ -80,4 +80,5 @@ public class JwtUtil {
         final String username = getUsernameFromToken(token);
         return (username.equals(userDetails.getUsername()) && !isTokenExpired(token));
     }
+
 }
